refactor: move BNVideoStream for his on file
